Refactored YouTubeChat. Now uses abstract ChatMessage-Format. Should be also compatible to twitch!
