+++ src/main/resources/sample.txt
@@ -1,18 +1,5 @@
 Andre;Hallo;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
-Andre;Ich heisse Andre;1
+*Andre;Ich heisse Andre;1
 Dennis;Hallo Andre;1
 Andre;red;1
 Andre;red;11000
