+++ src/main/scala/de/sebinside/codeoverflow/chatoverflow/backend/provider/twitch/TwitchChat.scala
@@ -1,0 +1,12 @@
+package de.sebinside.codeoverflow.chatoverflow.backend.provider.twitch
+
+import de.sebinside.codeoverflow.chatoverflow.backend.provider.MessageProvider
+
+/**
+  * Created by seb on 17.12.16.
+  */
+class TwitchChat extends MessageProvider {
+  override private[backend] def getMessages = ???
+
+  override private[backend] def getMessages(lastMilliseconds: Long) = ???
+}
