+++ src/main/scala/de/sebinside/codeoverflow/chatoverflow/backend/provider/youtube/Auth.java
@@ -1,0 +1,78 @@
+package de.sebinside.codeoverflow.chatoverflow.backend.provider.youtube;
+
+import com.google.api.client.auth.oauth2.Credential;
+import com.google.api.client.auth.oauth2.StoredCredential;
+import com.google.api.client.extensions.java6.auth.oauth2.AuthorizationCodeInstalledApp;
+import com.google.api.client.extensions.jetty.auth.oauth2.LocalServerReceiver;
+import com.google.api.client.googleapis.auth.oauth2.GoogleAuthorizationCodeFlow;
+import com.google.api.client.googleapis.auth.oauth2.GoogleClientSecrets;
+import com.google.api.client.http.HttpTransport;
+import com.google.api.client.http.javanet.NetHttpTransport;
+import com.google.api.client.json.JsonFactory;
+import com.google.api.client.json.jackson2.JacksonFactory;
+import com.google.api.client.util.store.DataStore;
+import com.google.api.client.util.store.FileDataStoreFactory;
+
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.Reader;
+import java.util.List;
+
+/**
+ * Shared class used by every sample. Contains methods for authorizing a user and caching credentials.
+ */
+class Auth {
+
+    /**
+     * Define a global instance of the HTTP transport.
+     */
+    protected static final HttpTransport HTTP_TRANSPORT = new NetHttpTransport();
+
+    /**
+     * Define a global instance of the JSON factory.
+     */
+    protected static final JsonFactory JSON_FACTORY = new JacksonFactory();
+
+    /**
+     * This is the directory that will be used under the user's home directory where OAuth tokens will be stored.
+     */
+    private static final String CREDENTIALS_DIRECTORY = ".oauth-credentials";
+
+    /**
+     * Authorizes the installed application to access user's protected data.
+     *
+     * @param scopes              list of scopes needed to run youtube upload.
+     * @param credentialDatastore name of the credential datastore to cache OAuth tokens
+     */
+    protected static Credential authorize(List<String> scopes, String credentialDatastore) throws IOException {
+
+        // Load client secrets.
+        Reader clientSecretReader = new InputStreamReader(Auth.class.getResourceAsStream("/client_secrets.json"));
+        GoogleClientSecrets clientSecrets = GoogleClientSecrets.load(JSON_FACTORY, clientSecretReader);
+
+        // Checks that the defaults have been replaced (Default = "Enter X here").
+        if (clientSecrets.getDetails().getClientId().startsWith("Enter")
+                || clientSecrets.getDetails().getClientSecret().startsWith("Enter ")) {
+            System.out.println(
+                    "Enter Client ID and Secret from https://console.developers.google.com/project/_/apiui/credential "
+                            + "into src/main/resources/client_secrets.json");
+            System.exit(1);
+        }
+
+        // This creates the credentials datastore at ~/.oauth-credentials/${credentialDatastore}
+        FileDataStoreFactory fileDataStoreFactory = new FileDataStoreFactory(new File(System.getProperty("user.home") + "/" + CREDENTIALS_DIRECTORY));
+        DataStore<StoredCredential> datastore = fileDataStoreFactory.getDataStore(credentialDatastore);
+
+        GoogleAuthorizationCodeFlow flow = new GoogleAuthorizationCodeFlow.Builder(
+                HTTP_TRANSPORT, JSON_FACTORY, clientSecrets, scopes).setCredentialDataStore(datastore)
+                .build();
+
+        // Build the local server and bind it to port 8080
+       LocalServerReceiver localReceiver = new LocalServerReceiver.Builder().setPort(8080).build();
+
+        // Authorize.
+       return new AuthorizationCodeInstalledApp(flow, localReceiver).authorize("user");
+
+    }
+}
